Work around inconsistent scandir declaration
Newer versions declare a comparison function prototype that uses typed
pointers, earlier versions use void.. or something. It might be that
earlier versions allow the PDK libc header to shadow the builtin one.
Anyway, there's a provided alphasort function that behaves close enough
to sortComparD to use instead, which is declared next to scandir. The
only difference is that alphasort uses strcoll instead of strcasecmp.