Cap auto frameskipping penalty for speeds below 80
Because the sound gets written on the main thread, when frameskipping is
high, the thread blocks waiting for the sound to get consumed, which
makes the autoframeskip code think things are slow during the next
iteration. The associated penalties result in an artificially high frame
skip setting, with lots of time wasting blocking on the sound
consumption. This change makes it easier for the frame skipping to be
reduced without compromising the responsiveness to slowdowns.