Pixi: Reduce spacing in the options and help screens
Preprocessor conditionals are used to reduce the spacing of options
items and help lines if PALM_PIXI is defined. This allows those screens
to fit into the Pixi's 320x400 screen resolution.

I think the smaller help line spacing should be the same on the Pre,
because it looks better. However, I don't want to impose that decision
on anyone, so I'm leaving it as is, for now.