Slip the frame skipping deadline to account for sdlBufferEmpty delay
This change eliminates the spurious frame skipping that happens when
sdlBufferEmpty is waited on. In exchange for this, there is a slowdown
in the emulation that causes some fairly annoying sound stutters. The
slowdown could be avoided by changing the frameskip code to spread the
delay penalty among multiple frames, but it would be better to fix the
sound synchronization instead.