Rewrite SDL Audio code without full and empty waiting
This change rewrites the sound output code to loosely synchronize
instead of using semaphores to wait for data to be ready. This means
that it's now possible for buffer underruns and overruns to occur, and
right now they occur too frequently for my taste. On the other hand,
this buys us smooth auto frame skipping code, eliminates the focus/audio
related freeze, which was a deadlock while waiting for sdlBufferEmpty,
and fixes the inconsistent sound output during speedup mode. Even if the
audio isn't perfect, vba is much more usable with this change than it
was previously.